+++ src/com/github/kingtim1/jmdp/discounted/MapVFunction.java
@@ -55,7 +55,7 @@ public class MapVFunction<S> implements DiscountedVFunction<S> {
 	}
 	
 	@Override
-	public double value(S state, Long timestep) {
+	public double value(S state, Integer timestep) {
 		return value(state);
 	}
